feat: Move checkConfigProperty to utils
